Refactoring to move exported API out of studioml runner.
