+++ studio/local_worker.py
@@ -15,8 +15,7 @@ from . import fs_tracker, model, logs
 from .local_queue import LocalQueue
 from .gpu_util import get_available_gpus, get_gpu_mapping, get_gpus_summary
 from .experiment import Experiment
-from .util import sixdecode, str2duration, retry, LogReprinter
-from .model import parse_verbosity
+from .util import sixdecode, str2duration, retry, LogReprinter, parse_verbosity
 
 logs.getLogger('apscheduler.scheduler').setLevel(logs.ERROR)
